# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AckRemindOut": {
        "description": "L4: how many members were reminded (those already reminded and still open are skipped).",
        "properties": {
          "reminded": {
            "title": "Reminded",
            "type": "integer"
          }
        },
        "required": [
          "reminded"
        ],
        "title": "AckRemindOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReminderCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "remind_at": {
            "format": "date-time",
            "title": "Remind At",
            "type": "string"
          }
        },
        "required": [
          "remind_at"
        ],
        "title": "ReminderCreate",
        "type": "object"
      },
      "ReminderOut": {
        "properties": {
          "channel_id": {
            "format": "uuid",
            "title": "Channel Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fired At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "personal",
            "enum": [
              "personal",
              "ack"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "preview": {
            "title": "Preview",
            "type": "string"
          },
          "remind_at": {
            "format": "date-time",
            "title": "Remind At",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "fired",
              "done",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "message_id",
          "channel_id",
          "note",
          "preview",
          "remind_at",
          "status",
          "fired_at",
          "created_at"
        ],
        "title": "ReminderOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ChikuwaChat API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/messages/{message_id}/ack/remind": {
      "post": {
        "description": "L4: the author (or an administrator) reminds the members who have not acknowledged;\neach gets a reminder only they see. Once an hour per message (429 ack_remind_too_soon).",
        "operationId": "remind_unacknowledged_api_v1_messages__message_id__ack_remind_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckRemindOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remind Unacknowledged",
        "tags": [
          "reminders"
        ]
      }
    },
...
    "/api/v1/messages/{message_id}/reminders": {
      "post": {
        "description": "M12e 「リマインド」: a nudge about this message at remind_at (at least a minute ahead).",
        "operationId": "create_reminder_api_v1_messages__message_id__reminders_post",
        "parameters": [
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reminder",
        "tags": [
          "reminders"
        ]
      }
    },
...
    "/api/v1/reminders": {
      "get": {
        "description": "My open reminders: fired ones first (newest nudge on top), then pending by time.",
        "operationId": "list_reminders_api_v1_reminders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderOut"
                  },
                  "title": "Response List Reminders Api V1 Reminders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reminders",
        "tags": [
          "reminders"
        ]
      }
    },
    "/api/v1/reminders/{reminder_id}": {
      "delete": {
        "description": "Cancels a pending reminder or marks a fired one done.",
        "operationId": "close_reminder_api_v1_reminders__reminder_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reminder_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reminder Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close Reminder",
        "tags": [
          "reminders"
        ]
      }
    },
...
  }
}